refactor: Move some utilities
